% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeLR.R
\name{cv_ridge_logistic_regression}
\alias{cv_ridge_logistic_regression}
\title{Find the optimal value for the tuning parameter of  ridge logistic regression
by k-folds cross validation.}
\usage{
cv_ridge_logistic_regression(y, X, lambdas, fold)
}
\arguments{
\item{y}{an n by 1 matrix containing the outcome variable}

\item{X}{an n by p design matrix}

\item{lambdas}{a sequence of values of tuning parameters}

\item{fold}{number of folds for cross validation}
}
\value{
The optimal tuning parameter for ridge logistic regression by
grid search
}
\description{
This function finds the optimal value for the tuning parameter of ridge
logistic regression by a given vector of outcome variable, a design matrix,
a user-specified sequence of choice of tuning parameter, and the number of
folds of the cross validation.
}
\examples{

cv_ridge_logistic_regression(y, X, 0:10, 5)

}
