% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeLR.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Fit the logistic regression model}
\usage{
logistic_regression(y, X)
}
\arguments{
\item{y}{an n by 1 matrix containing the outcome variable}

\item{X}{an n by p design matrix}
}
\value{
the maximum likelihood estimates of the regression coefficients
}
\description{
This function fits the logistic regression model
based on a given vector of outcome variable and a design matrix.
}
\examples{

logistic_regression(y, X)

}
