% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeLR.R
\name{matrix_X}
\alias{matrix_X}
\title{Data Pre-processing for the Design Matrix}
\usage{
matrix_X(X)
}
\arguments{
\item{X}{an n by p design matrix which needs pre-process}
}
\value{
the pre-processed design matrix
}
\description{
This function pre-processes the format of the design matrix,
specifically converts the categorical variables to the dummy variables
using reference cell coding
}
\examples{

matrix_X(X)

}
