% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeLR.R
\name{ridge_logistic_regression_beta}
\alias{ridge_logistic_regression_beta}
\title{Fit the ridge logistic regression model for a given tuning parameter.}
\usage{
ridge_logistic_regression_beta(y, X, lambda)
}
\arguments{
\item{y}{an n by 1 matrix containing the outcome variable}

\item{X}{an n by p design matrix}

\item{lambda}{a numeric value of the tuning parameter}
}
\value{
the maximum likelihood estimates of the regression coefficients
}
\description{
This function fits the logistic regression model
based on a given vector of outcome variable, a design matrix,
and a choice of tuning parameter.
}
\examples{

ridge_logistic_regression_beta(y, X, 0.1)

}
